Exit with non-zero value when integration test fails
Integration tests were run in loop and it didn't pass through non-zero
exit codes. It doesn't make much difference when executed by a human
because there are notifications about failed build. In CI server,
however, this is not the case and script should exit with non-zero value
to indicate the problem.

I didn't use `set -e` because it exits on the first error. In this case
I think it's more convenient to see which versions fail instead of
failing on first problem.